backend/requirements.txt
# Django & REST Framework
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
dj-database-url==2.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# AI & OpenAI
openai==1.12.0

# Reddit Integration
praw==7.7.1

# Notifications
python-telegram-bot==20.7
twilio==8.10.0

# Utilities
python-decouple==3.8
requests==2.31.0
beautifulsoup4==4.12.2

# Development
django-debug-toolbar==4.2.0 